+++ epstein_files/output/html/html_dir.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 
 DEFAULT_HTML_DIR = Path('docs')
-IMAGES_DIR = Path('doc_images')
+DOC_IMAGES_DIR = Path('doc_images')
+IMAGES_DIR = Path('images')
 
 
 class HtmlDir:
@@ -14,8 +15,8 @@ class HtmlDir:
 
     @classmethod
     def image_url(cls, filename: str) -> str:
-        return str(IMAGES_DIR.joinpath(filename))
+        return str(DOC_IMAGES_DIR.joinpath(filename))
 
     @classmethod
     def local_pic_path(cls, filename: str) -> Path:
-        return DEFAULT_HTML_DIR.joinpath(IMAGES_DIR, filename)
+        return DEFAULT_HTML_DIR.joinpath(DOC_IMAGES_DIR, filename)