+++ scripts/create_sample_html.py
@@ -13,6 +13,7 @@ from epstein_files.documents.document import Document
 from epstein_files.documents.documents.doc_list import DocList
 from epstein_files.documents.email import Email
 from epstein_files.documents.other_file import OtherFile
+from epstein_files.epstein_files import PICS
 from epstein_files.output.doc_printer import DocPrinter
 from epstein_files.output.html.builder import table_to_html, panel_to_div
 from epstein_files.output.site.sites import Site
@@ -24,7 +25,7 @@ from epstein_files.util.helpers.data_helpers import flatten
 from epstein_files.util.helpers.file_helper import open_file_or_url
 from epstein_files.util.logging import logger
 
-SAMPLE_SIZE = 4
+SAMPLE_SIZE = 2
 
 TEST_PANELS = [
     Panel('bright_red', style='bright_red'),
@@ -41,7 +42,7 @@ def print_sample_people(num_people_to_print: int = 3):
     good_sample_people = [
         p for
         p in epstein_files.emailers
-        if p.name != HOWARD_LUTNICK and (5 <= len(p.unique_emails) <= 15)
+        if p.name not in [HOWARD_LUTNICK, 'Roy Black'] and (5 <= len(p.unique_emails) <= 12)
     ]
 
     for i, person in enumerate(good_sample_people[0:num_people_to_print], 1):
@@ -55,22 +56,23 @@ def print_test_panels():
 
 
 doc_types_to_sample = [
-    # [d for d in epstein_files.documents if d.file_id == 'EFTA00009622'],
-    # [d for d in epstein_files.documents if d.category == 'article' and d._config.show_full_panel],
-    # [d for d in epstein_files.documents if d.category == 'government' and d._config.show_full_panel],
-    # [d for d in epstein_files.documents if d.category == 'money' and d._config.show_full_panel],
-    # [d for d in epstein_files.documents if d._config.pic_cfg],
-    # [d for d in epstein_files.documents if d._config.is_displayed_as_img],
-    # [d for d in epstein_files.documents if d._config.background_color], # Configured BG
-    # [d for d in epstein_files.other_files if d.category in CATEGORY_BG_STYLES],  # BG by category
-    # [e for e in epstein_files.emails if 'https' in e.text[0:1500]],
-    # [o for o in epstein_files.other_files if o.config and o.config.show_full_panel],
-    # [d for d in epstein_files._documents if d.suppressed_txt],
-    # [o for o in epstein_files.other_files if o._config.note_txt()],  # other file with description
-    # [o for o in epstein_files.other_files if 1000 < o.length < 5000 and not o._config.note_txt], # other files no desc
-    # [e for e in epstein_files.emails if e._config.note_txt()], # emails with description
-    # [e for e in epstein_files.emails if not e._config.note_txt()],  # email no desc
-    # epstein_files.emails_with_attachments,
+    PICS,
+    [d for d in epstein_files.documents if d.file_id == 'EFTA00009622'],
+    [d for d in epstein_files.documents if d.category == 'article' and d._config.show_full_panel],
+    [d for d in epstein_files.documents if d.category == 'government' and d._config.show_full_panel],
+    [d for d in epstein_files.documents if d.category == 'money' and d._config.show_full_panel],
+    [d for d in epstein_files.documents if d._config.pic_cfg],
+    [d for d in epstein_files.documents if d._config.is_displayed_as_img],
+    [d for d in epstein_files.documents if d._config.background_color], # Configured BG
+    [d for d in epstein_files.other_files if d.category in CATEGORY_BG_STYLES],  # BG by category
+    [e for e in epstein_files.emails if 'https' in e.text[0:1500]],
+    [o for o in epstein_files.other_files if o.config and o.config.show_full_panel],
+    [d for d in epstein_files._documents if d.suppressed_txt],
+    [o for o in epstein_files.other_files if o._config.note_txt()],  # other file with description
+    [o for o in epstein_files.other_files if 1000 < o.length < 5000 and not o._config.note_txt], # other files no desc
+    [e for e in epstein_files.emails if e._config.note_txt()], # emails with description
+    [e for e in epstein_files.emails if not e._config.note_txt()],  # email no desc
+    epstein_files.emails_with_attachments,
     epstein_files.imessage_logs,
 ]
 
@@ -79,14 +81,14 @@ sample_docs = DocList.uniquify_by_id(flatten([docs[:SAMPLE_SIZE] for docs in doc
 printer = DocPrinter(epstein_files=epstein_files)
 
 # print header
-printer.print_title_page_top()
-printer.print_title_page_bottom()
+# printer.print_title_page_top()
+# printer.print_title_page_bottom()
 
 # Print docs
-printer.print_documents(sample_docs)
+printer.print_documents(DocList.sort_by_timestamp(sample_docs))
 
 # print some People and their emails
-print_sample_people()
+# print_sample_people()
 
 #Print big emailers summary table
 # all_emailers = sorted(epstein_files.emailers, key=lambda person: person.sort_key)